# app/Http/Resources/ReportResource.php
<?php

namespace App\Http\Resources;

use Illuminate\Http\Request;
use Illuminate\Http\Resources\Json\JsonResource;

class ReportResource extends JsonResource
{
    /**
     * Transform the resource into an array.
     *
     * @return array<string, mixed>
     */
    public function toArray(Request $request): array
    {
        return [
            'worker_id' => $this->worker_id,
            'start_work' => $this->start_work,
            'end_work' => $this->end_work,
        ];
    }
}